'04/12/19'

+++ app/controllers/package.js
@@ -21,13 +21,16 @@ exports.create = async (req,res,next) => {
     } catch (error) {
         next(error);
     }
-}
+};
 
 exports.delete = async (req,res,next) => {
     const Package = require('../models/package');
     const id = req.params.id;
 
     try {
+        let name = await Package.find({'name':req.body.name}).exec(); //usuario de whm
+        if(name.length != 0) return res.status(400).json({ok:false,message:'Package already exists'});
+
         const package = await Package.findByIdAndDelete(id).exec();
         res.status(200).json({ok:true,package});
     } catch (error) {
@@ -37,12 +40,24 @@ exports.delete = async (req,res,next) => {
 
 exports.get = async (req,res,next) => {
     const Package = require('../models/package');
+    const count = await Package.estimatedDocumentCount().exec();
 
     try {
         let packages = await Package.find().exec();
-        res.status(200).json({ok:true,data:{results:packages}});
+        res.status(200).json({ok:true, data:{count,results:packages}});
     } catch (error) {
         next(error);
     }
-}
+};
 
+exports.search = async (req,res) => {
+    const Package = require('../models/package');
+    let value = new RegExp(req.params.value,'i');//usando regex i se transforma a minuscula
+    try {
+        let companies =  await Company.find({"name":{$in:value}}).exec();
+        let count = await Company.estimatedDocumentCount().exec();
+        res.status(200).json({ok:true, data:{count,results:companies}});
+    } catch (error) {
+        res.status(500).json({ok:false,error});
+    }
+};
